requirements.txt
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
daily-python==0.25.0
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
google==3.0.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud==0.34.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-storage==3.10.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
packaging==26.0
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vapi-python==0.1.9
